Change float to double

+++ CSSPModels/ClimateSite.cs
@@ -13,13 +13,13 @@ namespace CSSPModels
         public int ECDBID { get; set; }
         public string ClimateSiteName { get; set; }
         public string Province { get; set; }
-        public float? Elevation_m { get; set; }
+        public double? Elevation_m { get; set; }
         public string ClimateID { get; set; }
         public int? WMOID { get; set; }
         public string TCID { get; set; }
         public bool? IsProvincial { get; set; }
         public string ProvSiteID { get; set; }
-        public float? TimeOffset_hour { get; set; }
+        public double? TimeOffset_hour { get; set; }
         public string File_desc { get; set; }
         public DateTime? HourlyStartDate_Local { get; set; }
         public DateTime? HourlyEndDate_Local { get; set; }
